requirements.txt
flask
gunicorn
markdown
python-slugify
mdx_truly_sane_lists
